refactor to provide browser support at juice/client
