+++ lib/property.js
@@ -13,7 +13,7 @@ module.exports = exports = Property;
  * Module dependencies.
  */
 
-var compare = require('./utils').compare;
+var utils = require('./utils');
 
 /**
  * CSS property constructor.
@@ -39,7 +39,7 @@ function Property (prop, value, selector) {
 Property.prototype.compare = function (property) {
   var a = this.selector.specificity();
   var b = property.selector.specificity();
-  var winner = compare(a, b);
+  var winner = utils.compare(a, b);
 
   if (winner === a && a !== b) {
     return this;
