+++ lib/utils.js
@@ -8,7 +8,13 @@ var cssom = require('cssom');
 var cheerio = require('cheerio');
 var own = {}.hasOwnProperty;
 var os = require('os');
-var inline = require('web-resource-inliner');
+var Selector = require('./selector');
+var Property = require('./property');
+
+exports.Selector = Selector;
+exports.Property = Property;
+exports.styleSelector = new Selector('<style attribute>', [1, 0, 0, 0]);
+exports.importantSelector = new Selector('<!important>', [2, 0, 0, 0]);
 
 /**
  * Returns an array of the selectors.
@@ -110,7 +116,6 @@ exports.getPreservedText = function ( css, options )
 		  FONT_FACE: 5,
 		*/
 
-
         if ( options.fontFaces && rules[i].type === cssom.CSSFontFaceRule.prototype.type )
  		{
             var fontFace = [ "" ];
@@ -160,11 +165,6 @@ exports.cheerio = function (html, options) {
   options = exports.extend({decodeEntities: false}, options || {});
   html = exports.encodeEntities(html);
   return cheerio.load(html,options);
-};
-
-exports.inlineExternal = function (html, inlineOptions, callback){
-  var options = exports.extend({fileContent: html}, inlineOptions);
-  inline.html(options, callback);
 };
 
 exports.normalizeLineEndings = function (text){
@@ -234,3 +234,20 @@ exports.extend = function (obj, src) {
   }
   return obj;
 };
+
+exports.getDefaultOptions = function (options) {
+  var result = exports.extend({
+    extraCss: "",
+    applyStyleTags: true,
+    removeStyleTags: true,
+    preserveMediaQueries: false,
+    preserveFontFaces: false,
+    applyWidthAttributes: false,
+    applyAttributesTableElements: false,
+    url: ""
+  }, options);
+
+  result.webResources = result.webResources || {};
+
+  return result;
+}
